requirements.txt
aiohttp
asyncpg
aiohttp_debugtoolbar
aiohttp-swagger
pyyaml
uvloop
gunicorn
